extract validator to its own producer
