+++ server/src/main/java/org/zanata/magpie/service/MTStartup.java
@@ -32,7 +32,7 @@ import org.zanata.magpie.model.Role;
 import org.zanata.magpie.event.AccountCreated;
 import org.zanata.magpie.util.PasswordUtil;
 
-import static org.zanata.magpie.service.ResourceProducer.REPLICATE_CACHE;
+import static org.zanata.magpie.producer.ResourceProducer.REPLICATE_CACHE;
 
 /**
  * Startup monitor for MT.
